Entry for log code in rounds_detail updated to spinner.
