+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".roundsDetail"
+        <activity android:name=".RoundsDetail"
             android:exported="true"></activity>
     </application>
